fixed int/double issue

+++ src/main/java/org/alfresco/extensions/bulkexport/controler/Engine.java
@@ -188,7 +188,7 @@ public class Engine {
         List<Future<?>> futures = new ArrayList<>();
 
         int previousLowerLimitNodeNumber = 0;
-        int noOfTasks = new Double(Math.ceil(nodesToExport.size() / this.exportChunkSize)).intValue();
+        int noOfTasks = new Double(Math.ceil((double) nodesToExport.size() / (double) this.exportChunkSize)).intValue();
 
         log.info("Number of tasks: " + noOfTasks);
